FIX #361 - World object built-in
Adds a built-in object `world` to PGZero to allow easier
management of games that exceed one screen of space.

Add new actors to the world with `world.add_actor(a)` and then
draw all actors in the world at their positions with `world.draw()`.

To change the "camera view" of the world, shift its offsets with
`world.x` and `world.y`.