Add pixel perfect collision for actors
Adds `actor1.collidemask(actor2)` as a function to check pixel
perfect collisions.

The function first checks for collision via `.colliderect()` and
masks are only created for actors when necessary, meaning there
is no performance cost if the function isn't used in a project.

Adds documentation and tests. The documentation makes it clear
that this function has a bigger impact on performance than
`.colliderect()` and should only be used sparingly.